requirements.txt
flask
flask-login
flask-sqlalchemy
torch
torchvision
transformers
accelerate
numpy
pillow
opencv-python
reportlab